Split classes to utils
